Backend/requirements.txt
aiohttp==3.8.6
aiosignal==1.3.1
alabaster==0.7.13
altgraph==0.17.3
annotated-types==0.6.0
anyio==4.2.0
arabic-reshaper==3.0.0
arrow==1.2.3
asgiref==3.7.2
asn1crypto==1.5.1
astroid==2.15.6
asttokens==2.4.0
async-timeout==4.0.3
atomicwrites==1.4.1
attrs==23.1.0
autopep8==2.0.4
Babel==2.14.0
backcall==0.2.0
bcrypt==3.2.0
beautifulsoup4==4.12.2
bidict==0.22.1
binaryornot==0.4.4
black==23.7.0
bleach==6.0.0
blinker==1.7.0
certifi==2023.11.17
cffi==1.16.0
chardet==5.2.0
charset-normalizer==3.3.2
click==8.1.7
cloudpickle==2.2.1
colorama==0.4.6
comm==0.1.4
contourpy==1.2.0
cookiecutter==2.3.0
copier==8.3.0
cryptography==41.0.3
cssselect2==0.7.0
customtkinter==5.2.0
cycler==0.12.1
darkdetect==0.8.0
debugpy==1.7.0
decorator==5.1.1
defusedxml==0.7.1
diff-match-patch==20230430
dill==0.3.7
distlib==0.3.7
Django==4.1.13
djongo==1.3.6
dnspython==2.5.0
docstring-to-markdown==0.12
docutils==0.20.1
dunamai==1.18.0
email-validator==2.1.0.post1
et-xmlfile==1.1.0
eventlet==0.35.0
exceptiongroup==1.2.0
executing==1.2.0
fastapi==0.108.0
fastjsonschema==2.18.0
filelock==3.13.1
flake8==6.0.0
Flask==3.0.2
flask-babel==4.0.0
Flask-BabelEx==0.9.4
Flask-Bcrypt==1.0.1
Flask-Cors==4.0.0
Flask-JWT-Extended==4.6.0
Flask-Login==0.6.3
Flask-Mail==0.9.1
flask-mongoengine==1.0.0
Flask-Principal==0.4.0
Flask-Security==3.0.0
Flask-Security-Too==5.3.3
Flask-SocketIO==5.3.6
Flask-WTF==1.2.1
flet==0.10.1
flet-core==0.10.1
flet-runtime==0.10.1
fonttools==4.47.0
freetype-py==2.3.0
frozenlist==1.4.0
funcy==2.0
greenlet==3.0.3
h11==0.14.0
html5lib==1.1
httpcore==0.17.3
httptools==0.6.1
httpx==0.24.1
idna==3.6
imagesize==1.4.1
importlib-metadata==6.8.0
importlib-resources==6.1.1
inflection==0.5.1
iniconfig==2.0.0
intervaltree==3.1.0
ipykernel==6.25.2
ipython==8.15.0
ipython-genutils==0.2.0
isort==5.12.0
itsdangerous==2.1.2
jaraco.classes==3.3.0
jedi==0.18.2
jellyfish==1.0.0
Jinja2==3.1.3
jinja2-ansible-filters==1.3.2
jsonschema==4.19.0
jsonschema-specifications==2023.7.1
jupyter_client==8.3.1
jupyter_core==5.3.1
jupyterlab-pygments==0.2.2
keyring==24.2.0
kiwisolver==1.4.5
lazy-object-proxy==1.9.0
lxml==4.9.3
markdown-it-py==3.0.0
MarkupSafe==2.1.4
matplotlib==3.8.2
matplotlib-inline==0.1.6
mccabe==0.7.0
mdurl==0.1.2
mistune==3.0.1
mongoengine==0.27.0
more-itertools==10.1.0
motor==3.3.2
mpmath==1.3.0
multidict==6.0.4
mypy-extensions==1.0.0
mysql-connector-python==8.0.33
mysqlclient==2.2.0
nbformat==5.9.2
nest-asyncio==1.5.7
numpy==1.26.2
oauthlib==3.2.2
openai==0.28.1
openpyxl==3.1.2
oscrypto==1.3.0
packaging==23.2
pandas==2.1.3
pandocfilters==1.5.0
paramiko==3.3.1
parso==0.8.3
passlib==1.7.4
pathspec==0.11.2
pefile==2023.2.7
pexpect==4.8.0
pickleshare==0.7.5
Pillow==10.1.0
platformdirs==4.0.0
pluggy==1.3.0
plumbum==1.8.2
prompt-toolkit==3.0.36
protobuf==3.20.3
psutil==5.9.5
ptyprocess==0.7.0
pure-eval==0.2.2
pycairo==1.25.1
pycodestyle==2.10.0
pycparser==2.21
pydantic==2.5.3
pydantic_core==2.14.6
pydocstyle==6.3.0
pyflakes==3.0.1
Pygments==2.16.1
pyHanko==0.20.1
pyhanko-certvalidator==0.24.1
pyinstaller==5.13.2
pyinstaller-hooks-contrib==2023.8
PyJWT==2.8.0
pylint==2.17.5
pylint-venv==3.0.2
pyls-spyder==0.4.0
pymongo==4.6.1
PyNaCl==1.5.0
pyparsing==3.1.1
pypdf==3.17.1
pypng==0.20220715.0
PyQt5==5.15.9
PyQt5-Qt5==5.15.2
PyQt5-sip==12.12.2
PyQtWebEngine==5.15.6
PyQtWebEngine-Qt5==5.15.2
pytest==7.4.3
python-bidi==0.4.2
python-dateutil==2.8.2
python-decouple==3.8
python-dotenv==1.0.0
python-engineio==4.8.2
python-lsp-black==1.3.0
python-lsp-jsonrpc==1.0.0
python-lsp-server==1.7.4
python-slugify==8.0.1
python-socketio==5.11.0
pytoolconfig==1.2.5
pytz==2024.1
pywin32==306
pywin32-ctypes==0.2.2
PyYAML==6.0.1
pyyaml-include==1.3.1
pyzmq==25.1.1
QDarkStyle==3.1
qrcode==7.4.2
qstylizer==0.2.2
QtAwesome==1.2.3
qtconsole==5.4.4
QtPy==2.4.0
questionary==2.0.1
referencing==0.30.2
repath==0.9.0
reportlab==4.0.7
requests==2.31.0
rich==13.5.2
rlPyCairo==0.3.0
rope==1.9.0
rpds-py==0.10.2
Rtree==1.0.1
simple-websocket==1.0.0
six==1.16.0
sniffio==1.3.0
snowballstemmer==2.2.0
sortedcontainers==2.4.0
soupsieve==2.5
speaklater==1.3
sphinxcontrib-jsmath==1.0.1
spyder-kernels==2.4.4
sqlparse==0.2.4
stack-data==0.6.2
starlette==0.32.0.post1
style==1.1.0
svglib==1.5.1
sympy==1.12
text-unidecode==1.3
textdistance==4.5.0
three-merge==0.1.1
tinycss2==1.2.1
tomli==2.0.1
tomlkit==0.12.1
tornado==6.3.3
tqdm==4.66.1
traitlets==5.9.0
ttkthemes==3.2.2
typing_extensions==4.9.0
tzdata==2023.3
tzlocal==5.2
ujson==5.8.0
update==0.0.1
uritools==4.0.2
urllib3==2.1.0
uvicorn==0.25.0
virtualenv==20.24.7
watchdog==3.0.0
watchfiles==0.21.0
wcwidth==0.2.6
webencodings==0.5.1
websocket-client==1.6.3
websockets==12.0
Werkzeug==3.0.1
whatthepatch==1.0.5
wrapt==1.15.0
wsproto==1.2.0
WTForms==3.1.2
xhtml2pdf==0.2.13
yapf==0.40.1
yarl==1.9.2
zipp==3.16.2
